Move utils/ api.ts and query.ts to utils/api/
