+++ apps/client/src/features/onboarding/api/options.ts
@@ -1,6 +1,6 @@
 import { mutationOptions } from "@tanstack/react-query"
 
-import { client } from "@/utils/api"
+import { client } from "@/utils/api/client"
 import { delayWithPromise } from "@/utils/delay"
 
 export const onboardingMutationOptions = (minimumDelayMs: number) =>
